requirements.txt
streamlit
openai
duckduckgo-search
python-dotenv
csv
os